backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg[binary]==3.3.2
python-decouple==3.8
django-cors-headers==4.3.1
drf-spectacular==0.27.0
django-jazzmin==2.6.0
